main/requirements.txt
tensorflow
numpy
Pillow
matplotlib
requests
opencv-python